refactor(graph): split frontmatter->RDF converter out of indexers.ts (#1905)
indexers.ts:111-401 was one cohesive value-mapping concern that never
touched orchestration: recoverYamlEatenWikiLink, isFrontmatterMap,
emitFrontmatterValue, resolveFrontmatterPredicate, declaredPropertyPredicate,
the ISO/wikilink regexes, resolveWholeWikiLink, frontmatterValueToEdge,
coerceDeclared. The indexers/ directory already held 4 sibling modules
(excerpt, source, note-files, tables) split out the same way — this one was
half-done.

Moves the block to indexers/frontmatter.ts, matching the sibling modules'
plain-function style (no Ops-closure needed here — this is pure backend
value mapping, not component state). reloadTypeCatalog stayed behind: it sat
in the middle of the old block by coincidence but is type-catalog
orchestration, not frontmatter value mapping, and isn't in the issue's
function list.

indexers.ts re-exports resolveFrontmatterPredicate and
declaredPropertyPredicate so note-properties.ts's `import { … } from
'./indexers'` is unchanged — same "the barrel stays stable" pattern this
file's own header comment already documents for graph/index.ts.

992 -> 723 lines. BUDGETS entry lowered in the same diff. No coverage-floor
retune needed: indexers/frontmatter.ts falls under the existing
`src/main/graph/**` aggregate threshold, and moving already-well-tested code
between files under the same aggregate doesn't change its ratio. The three
frontmatter test files named in the issue (frontmatter-indexing,
frontmatter-backlinks, frontmatter-link-parity) pass unchanged.
